+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - **[UPDATE]** Don't deactivate/activate the trap in `useFocusTrap` each time `onClose` changes.
+- **[UPDATE]** Add `dataEncryptedName` prop on `TextField`
 - [...]
 
 # v41.7.1 (17/11/2020)